Add first version of results class
Add first version of results class together with unit tests. This
version includes a parent class and a child class for TE results. Child
classes for further network analysis algorithms will be added in future
commits.

Methods for combining results from idtxl_utils are moved to the
results class. Instead, functions for testing dictionaries for
conflicting entries or for equality are added to the utils module.

Methods for constructing adjacency matrices and reconstructing delays
are moved from visualise_graph to the results class.